% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPipeline.R
\name{graphPipeline}
\alias{graphPipeline}
\title{Convert a pipeline input/output data frame into an igraph object.}
\usage{
graphPipeline(pipeline_dataframe)
}
\arguments{
\item{pipeline_dataframe}{An input/output data frame returned by \code{\link[=mapPipeline]{mapPipeline()}}.}
}
\value{
A directional igraph object describing the workflow of the mapped
pipeline directory.
}
\description{
Convert a pipeline input/output data frame into an igraph object.
}
\examples{
example_directory <- system.file("dummy_pipeline", package = "pipelinemapper")
pipeline_dataframe <- mapPipeline(example_directory)
graphPipeline(pipeline_dataframe)


}
