% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPipeline.R
\name{mapPipeline}
\alias{mapPipeline}
\title{A wrapper function that applies mapScript() to target files within a directory.}
\usage{
mapPipeline(
  pipeline_directory_path,
  file_extensions = c(".r", ".rmd", ".qmd"),
  input_tag = "#input",
  output_tag = "#output"
)
}
\arguments{
\item{pipeline_directory_path}{The path to the pipeline directory.}

\item{file_extensions}{A character vector containing all file extensions to be considered.}

\item{input_tag}{A character tag for input paths within the target pipeline.}

\item{output_tag}{A character tag for output paths within the target pipeline.}
}
\value{
A data frame containing columns describing the target
pipeline's input and and output paths. See \code{\link[=mapScript]{mapScript()}}.
}
\description{
This script loops \code{\link[=mapScript]{mapScript()}} through all scripts within a target directory
containing the specified file extensions. The input and output paths are compiled
into a single output data frame.
}
\examples{
example_directory <-
  system.file("dummy_pipeline",
              package = "pipelinemapper")

mapPipeline(pipeline_directory_path = example_directory,
           file_extensions = c(".r",".rmd",".qmd"),
           input_tag = "#input",
           output_tag = "#output")

}
