% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPipeline.R
\name{plotPipeline}
\alias{plotPipeline}
\title{Plots a pipeline graph}
\usage{
plotPipeline(pipeline_graph)
}
\arguments{
\item{pipeline_graph}{A pipeline igraph object returned by \code{\link[=graphPipeline]{graphPipeline()}}.}
}
\value{
A ggraph plot object
}
\description{
This is a default way of plotting the igraph object returned by \code{\link[=graphPipeline]{graphPipeline()}}
using ggraph and ggplot2. If you would like to customise the plot, it is
probably easier to refer directly to igraph, ggraph, and ggplot2
documentation.
}
\examples{
1 + 1

}
