% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPipeline.R
\name{mapPipeline}
\alias{mapPipeline}
\title{Generates a data frame of file inputs and outputs for scripts within a pipeline directory.}
\usage{
mapPipeline(pipeline_directory_path, file_extensions = c(".r", ".rmd", ".qmd"))
}
\arguments{
\item{pipeline_directory_path}{The path to the pipeline directory.}

\item{file_extensions}{A character vector containing all file extensions to be considered.}
}
\value{
A data frame containing five columns: input/output file name,
input/output path, script name, script path, and direction (in or out).
}
\description{
This script loops \code{\link[=mapScript]{mapScript()}} through all scripts within a target directory
containing the specified file extensions. The inputs and outputs are compiled
into a single output data frame.
}
\examples{
example_directory <- system.file("dummy_pipeline", package = "pipelinemapper")
mapPipeline(example_directory)

}
