% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapScript.R
\name{mapScript}
\alias{mapScript}
\title{Extract input and output files from a pipeline script.}
\usage{
mapScript(script_path, input_tag = "#input", output_tag = "#output")
}
\arguments{
\item{script_path}{The path to the target script as a character string.}

\item{input_tag}{A character tag for input paths within the target script.}

\item{output_tag}{A character tag for output paths within the target script.}
}
\value{
A data frame containing five columns: input/output file name, input/output path, script name, script path, and direction ("in" or "out")
}
\description{
Extract input and output files from a pipeline script.
}
\examples{
example_script <- system.file("dummy_pipeline/01_load_data.R", package = "pipelinemapper")
mapScript(example_script,input_tag  = "#input",output_tag = "#output")

}
