% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPipeline.R
\name{plotPipeline}
\alias{plotPipeline}
\title{Plots a pipeline graph}
\usage{
plotPipeline(pipeline_graph, show_full_paths = TRUE)
}
\arguments{
\item{pipeline_graph}{A pipeline igraph object returned by \code{\link[=graphPipeline]{graphPipeline()}}.}

\item{show_full_paths}{Whether or not full file paths should be displayed.
show_full_paths = FALSE only displays the file basenames.}
}
\value{
A ggraph plot object
}
\description{
This is a default way of plotting the igraph object returned by \code{\link[=graphPipeline]{graphPipeline()}}
using ggraph and ggplot2. If you would like to customise the plot, it is
probably easier to refer directly to igraph, ggraph, and ggplot2
documentation.
}
\examples{
example_directory <- system.file("dummy_pipeline", package = "pipelinemapper")
pipeline_dataframe <- mapPipeline(example_directory)
pipeline_graph <- graphPipeline(pipeline_dataframe)
plotPipeline(pipeline_graph, show_full_paths = TRUE)

}
